# internal/role-leader.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "role-leader.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "RoleLeader"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/create-feedback": {
      "post": {
        "operationId": "RoleLeader_CreateFeedback",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreateFeedbackResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateFeedbackRequest"
            }
          }
        ],
        "tags": [
          "RoleLeader"
        ]
      }
    },
    "/api/get-call/{callId}": {
      "get": {
        "operationId": "RoleLeader_GetCall",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetCallResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "callId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "RoleLeader"
        ]
      }
    },
    "/api/leaders-calls/{leaderId}": {
      "get": {
        "operationId": "RoleLeader_GetLeaderCalls",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetLeaderCallsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "leaderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "RoleLeader"
        ]
      }
    }
  },
  "definitions": {
    "apiCall": {
      "type": "object",
      "properties": {
        "callId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "leaderId": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "feedback": {
          "type": "string"
        },
        "startTime": {
          "type": "string"
        }
      }
    },
    "apiCreateFeedbackRequest": {
      "type": "object",
      "properties": {
        "callId": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "apiCreateFeedbackResponse": {
      "type": "object"
    },
    "apiGetCallResponse": {
      "type": "object",
      "properties": {
        "call": {
          "$ref": "#/definitions/apiCall"
        }
      }
    },
    "apiGetLeaderCallsResponse": {
      "type": "object",
      "properties": {
        "calls": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiCall"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}